% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAGR.R
\name{CAGR}
\alias{CAGR}
\title{Compound Annual Growth Rate}
\usage{
CAGR(start, end, periods)
}
\arguments{
\item{start}{starting value of an investment}

\item{end}{end value of an investment}

\item{periods}{number of periods covered during}
}
\description{
Calculate the compound annual growth rate for a given investment.
}
