% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateConv.R
\name{DateConv}
\alias{DateConv}
\title{Convert string to data}
\usage{
DateConv(x)
}
\arguments{
\item{x}{character string to convert to date value}
}
\description{
Converts LendingClubs data format for use in R
}
