% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FillNA.R
\name{FillNA}
\alias{FillNA}
\title{Fill NA values}
\usage{
FillNA(x, y)
}
\arguments{
\item{x}{value to check for NA}

\item{y}{value to plug in place if x is NA}
}
\description{
Fills NA values with the given plug value
}
