% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IssuedLoans.R
\name{IssuedLoans}
\alias{IssuedLoans}
\title{Dataset of 1.2MM peer-to-peer loans}
\usage{
IssuedLoans()
}
\value{
data frame of historical loans
}
\description{
A dataset of historical loan data including the loan status and over 100
other variables. Lending Club publishes this data on all loans it has issued
since 2007 and refreshes it quarterly. The data is published approximately
6 weeks after the quarter.
}
