% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LendingClubData.R
\docType{package}
\name{LendingClubData}
\alias{LendingClubData}
\alias{LendingClubData-package}
\title{Loan Data}
\description{
Historical loan data downloaded from Lending Club.  The data includes
over 100 variables for each loan dating back to 2007.
}
