% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LendingClubData.R
\docType{data}
\name{LendingClubData_2}
\alias{LendingClubData_2}
\title{Loan Data subset 2
Second subset of loans issued by Lending Club.  The dataset was
split into 2 data files to circumvent GitHub's file size limit.
To get the complete dataset, please use the IssuedLoans()
function.}
\description{
Loan Data subset 2
Second subset of loans issued by Lending Club.  The dataset was
split into 2 data files to circumvent GitHub's file size limit.
To get the complete dataset, please use the IssuedLoans()
function.
}
\keyword{data}
