% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCT.R
\name{PCT}
\alias{PCT}
\title{Round to 2 decimal places}
\usage{
PCT(x)
}
\arguments{
\item{x}{decimal value}
}
\description{
Simple wrapper for the round function which converts a decimal percentage
to a integer value
}
