% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepLoadData.R
\name{Prep_Loan_Data}
\alias{Prep_Loan_Data}
\title{Prep raw data download}
\usage{
Prep_Loan_Data(x)
}
\arguments{
\item{x}{file name of a file to convert.}
}
\description{
Takes a raw downloaded file and converts the columns to the correct data
types. The downloaded file includes character representation of numerical
values. For example, "25%" is converted to a numeric 25.
}
