% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetClass.R
\name{SetClass}
\alias{SetClass}
\title{Set the class of the loan}
\usage{
SetClass(x)
}
\arguments{
\item{x}{loan status}
}
\value{
character string. Set 'A' for current loans and
'B'  for loans that are late or in default status.
}
\description{
Set the class of the loan based on the loan status.
}
