% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_LC.R
\name{theme_LC}
\alias{theme_LC}
\title{Theme for charts in Lending Club ecosystem}
\usage{
theme_LC(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base text family}
}
\description{
Theme for charts in Lending Club ecosystem
}
