% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLoanData.R
\name{GetLoanData}
\alias{GetLoanData}
\title{Get loan data from Lending Club}
\usage{
GetLoanData(address, directory)
}
\arguments{
\item{address}{URL for data to download}

\item{directory}{relative path where the files will be saved}
}
\description{
Downloads the loan data from Lending Club's website. The data
downloads in zip format which this function will unzip.
}
